trading_bot/src/trading/trading_logic.py
import logging
from trading_bot.src.api.api_connector import APIConnector
from trading_bot.src.analysis.technical_analysis import TechnicalAnalysis
from trading_bot.src.analysis.qualitative_analysis import QualitativeAnalysis

class TradingLogic:
    def __init__(self, config):
        self.config = config
        self.api_connector = APIConnector(
            config['api']['base_url'],
            config['api']['api_key'],
            config['api']['api_secret'],
            config['api']['reconnect_attempts'],
            config['api']['reconnect_backoff']
        )
        self.technical_analysis = TechnicalAnalysis()
        self.qualitative_analysis = QualitativeAnalysis(
            config['qualitative_analysis']['google_news_api_key'],
            config['qualitative_analysis']['twitter_api_key'],
            config['qualitative_analysis']['twitter_api_secret']
        )
        self.logger = logging.getLogger(__name__)

    def execute_trade(self, symbol, order_type, quantity, price=None):
        """
        Execute a trade by placing an order through the API connector.

        :param symbol: The stock symbol to trade
        :param order_type: The type of order (market, limit, stop)
        :param quantity: The number of shares to trade
        :param price: The price for limit/stop orders (optional)
        """
        if not self.api_connector.is_connected():
            self.api_connector.connect()

        # Validate order parameters
        if quantity <= 0:
            self.logger.error("Quantity must be greater than zero")
            return
        if order_type in ["limit", "stop"] and price is None:
            self.logger.error("Price must be specified for limit/stop orders")
            return

        trade = self.api_connector.place_order(symbol, order_type, quantity, price)
        if trade:
            self.logger.info(f"Trade executed: {trade}")
        else:
            self.logger.error("Trade execution failed")

    def evaluate_trading_opportunity(self, symbol):
        """
        Evaluate trading opportunity by combining technical and qualitative analysis.

        :param symbol: The stock symbol to evaluate
        :return: Combined signal score
        """
        market_data = self.api_connector.get_market_data(symbol)
        if not market_data:
            self.logger.error("Failed to retrieve market data")
            return None

        technical_signal = self.technical_analysis.evaluate(market_data)
        qualitative_signal = self.qualitative_analysis.get_qualitative_analysis(symbol)

        if technical_signal and qualitative_signal:
            combined_signal = (technical_signal + qualitative_signal) / 2
            return combined_signal
        return None

    def manage_risk(self, symbol, position_size, current_price):
        """
        Manage risk by enforcing risk management rules.

        :param symbol: The stock symbol to trade
        :param position_size: The size of the position
        :param current_price: The current price of the stock
        :return: True if risk management rules are satisfied, False otherwise
        """
        max_position_size = self.config['trading']['max_position_size']
        daily_loss_limit = self.config['trading']['daily_loss_limit']
        max_trade_frequency = self.config['trading']['max_trade_frequency']

        if position_size > max_position_size:
            self.logger.warning(f"Position size for {symbol} exceeds maximum limit")
            return False

        if self.calculate_daily_loss() > daily_loss_limit:
            self.logger.warning("Daily loss limit exceeded")
            return False

        if self.calculate_trade_frequency(symbol) > max_trade_frequency:
            self.logger.warning("Maximum trade frequency exceeded")
            return False

        return True

    def calculate_daily_loss(self):
        """
        Calculate the daily loss.

        :return: Daily loss amount
        """
        # Placeholder for daily loss calculation logic
        return 0

    def calculate_trade_frequency(self, symbol):
        """
        Calculate the trade frequency for a given symbol.

        :param symbol: The stock symbol to calculate trade frequency for
        :return: Trade frequency
        """
        # Placeholder for trade frequency calculation logic
        return 0
